+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Retrieve PSR-7 request object from PSR-14 events via getRequest() method (#11)
 
+### Fixed
+- Initialisation of JavaScript variable _paq is now more robust (#13)
+
 ## [1.4.0] - 2022-10-05
 
 ### Added