+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.1] - 2023-04-01
+
 ### Fixed
 - Error when opening a site configuration in TYPO3 v12.3 (#15)
 
@@ -59,7 +61,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release
 
 
-[Unreleased]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.5.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.5.1...HEAD
+[1.5.1]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.3.2...v1.4.0
 [1.3.2]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.3.1...v1.3.2